proyect/requirements.txt
Django
gunicorn
django-cors-headers
djangorestframework
antlr4-tools
antlr4-python3-runtime==4.13.2